refactor(personas): extrair modelos de dados para _persona_models.py
Move a camada de modelos de dados (4 enums, 4 dataclasses, 1 schema Pydantic)
de `personas/base.py` para o novo módulo-folha `personas/_persona_models.py`.
O `base.py` passa a conter apenas as ABCs BasePersona/BaseAutonomousPersona
mais um shim de re-export verbatim dos 9 modelos, mantendo a superfície
pública 100% estável sem alterar nenhum importador existente.

Motivação: separação de concerns — módulo-folha sem import interno do DEILE
elimina por construção qualquer risco de ciclo. Mesmo arquétipo do loop 02
(_plan_models.py) e 03 (_coercions.py), já mergeados.

Delta declarado: 2 log records da camada de modelos (base.py:152 warning,
:507 debug) passam a emitir sob `deile.personas._persona_models` em vez de
`deile.personas.base`. Nenhum teste assere sobre esses nomes de logger.

Refs #754
